AI_Agents/requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
ddgs>=7.0.0
duckduckgo-search>=7.0.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
tavily-python>=0.5.0
